requirements.txt
amqp==5.1.1
arabic-reshaper==2.1.4
asgiref==3.6.0
asn1crypto==1.5.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cron-validator==1.0.6
cryptography==38.0.4
cssselect2==0.7.0
DateTime==4.9
Django==4.1.4
django-background-task==0.1.8
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-colorfield==0.5.0
django-compat==1.0.15
django-cors-headers==3.13.0
django-crontab==0.7.1
django-dotenv==1.4.2
django-enumchoicefield==2.0.0
django-environ==0.7.0
django-mssql-backend==2.8.1
django-phonenumber-field==7.0.1
django-phonenumbers==1.0.1
django-pyodbc==1.1.3
django-pyodbc-azure==1.2.5
django-rest-swagger==2.2.0
django-timezone-field==5.0
django-xhtml2pdf==0.0.4
django_debug_toolbar==3.8.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
dnspython==2.2.1
docopt==0.6.2
drf-yasg==1.21.4
environs==9.5.0
et-xmlfile==1.1.0
eventlet==0.33.2
future==0.18.2
gevent==22.10.2
greenlet==2.0.1
html5lib==1.1
idna==3.4
importlib-metadata==5.1.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
lxml==4.9.2
MarkupSafe==2.1.1
marshmallow==3.19.0
mssql-django==1.2
numpy==1.23.5
openapi-codec==1.3.2
openpyxl==3.0.10
oscrypto==1.3.0
packaging==22.0
pandas==1.5.2
phonenumbers==8.13.3
Pillow==9.3.0
pipreqs==0.4.11
prompt-toolkit==3.0.36
pycparser==2.21
pyHanko==0.15.1
pyhanko-certvalidator==0.19.8
PyJWT==2.6.0
pyodbc==4.0.35
PyPDF3==1.0.6
python-bidi==0.4.2
python-crontab==2.7.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
redis==4.4.0
reportlab==3.6.12
requests==2.28.1
ruamel.yaml==0.17.21
simplejson==3.18.0
six==1.16.0
sqlparse==0.4.3
svglib==1.4.1
tinycss2==1.2.1
tqdm==4.64.1
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
xhtml2pdf==0.2.8
yarg==0.1.9
zipp==3.11.0
zope.event==4.6
zope.interface==5.5.2
